Reformulate how we do clusters.
